Allow for kinematic replication in async mode. When async, the replication callback was always setting object state to dynamic even when it should be kinematic, which conflicted with GT trying to force it to be kinematic. This interfered with marshaling of position and velocity data from PT to GT.
Also make UpdateReplicationMode take an argument for whether or not it should behave as if the body is asleep. Unfortunately, ordering matters a lot - when dagwood's WakeUp is called, it needs to do UpdateReplicationState first to make sure it becomes either dynamic or kinematic BEFORE waking up. Then calls the Super::WakeUp to do the rest.

#rb michael.lentine

#ROBOMERGE-OWNER: steven.barnett
#ROBOMERGE-AUTHOR: steven.barnett
#ROBOMERGE-SOURCE: CL 15466441 via CL 15466863 via CL 15466900 via CL 15466914 via CL 15466936
#ROBOMERGE-BOT: RELEASE (Release-Engine-Staging -> Main) (v771-15082668)

[CL 15496516 by steven barnett in Main branch]