New skydome importance sampling for path tracer
Change the skydome importance sampling table used in the path tracer to use an equi-area mapping instead of cube maps. This simplifies the importance sampling step, preserves stratification better, and gives us an opportunity to sample the correct sky cubemap. The previous code only worked in static cases and did not support the real time capture mode or blending.

Currently this importance table is rebuilt every frame (by the path tracer), but this can be improved in a subsequent step. The skylight prep is relatively cheap (compared to the main path tracer kernel at least). For a 1k cubemap (which builds a 2k importance table) this pass costs roughly 0.4ms on a 2080 Ti.

The old cubemap based approach is still around for now to keep RT Skylight working as-is, but I'm hoping to share the code in a follow-up step.

Also disable the special case making the skylight always visible to camera. The new behavior better matches the raster path.

#rb Patrick.Kelly

[CL 15534666 by chris kulla in ue5-main branch]