Improve error-distribution of Sobol sampling
Provide an alternative entry point to initialize random sequencss that takes the target number of samples into account as well as the pixel coordinates. By sharing a single sequence among pixels, shuffled along a space filling curve the visual distribution of error is reduced. Owen scrambling is still used to hide the structure of this space filling curve.

This is currently off by default, but can be enabled with the cvar: "r.PathTracing.UseErrorDiffusion=1"

#rb Patrick.Kelly

#ushell-cherrypick of 14829446 by chris.kulla

[CL 15576998 by chris kulla in Main branch]