Batched Jupiter queries
Adds a helper class to make CachedDataProbablyExists, GetCachedData and PutCachedData operations. The helper checks the availble connections in the pool and decides if the thread should make a regular request or batch together with other requests. This will reduce the number of concurrent connections on each client.

Disabled by default.

#rb devin.doucette

[CL 15604924 by Johan Berg in ue5-main branch]