Control Rig: Added delegate for when a control rig is bound and unbound.
The Control Rig Actor has an internal pointer to a ControlRig based on the bindings bound that match the actor to track and the control rig class. This is established on Refresh() whenever certain properties are toggled.  However, it's not cleared and rebuilt when the bindings chang. A bug manifested in which control rig gizmos would appear when opening Sequencer but would not disappear when closing Sequencer. Sequencer is setting visibility and restoring visibility properly. However, because the bindings change on close, the control rig gizmos weren't getting cleared and refreshed.

#jira UE-106936
#rb mike.zyracki, helge.mathee

#ROBOMERGE-SOURCE: CL 15628192 via CL 15628193 via CL 15628198
#ROBOMERGE-BOT: RELEASE (Release-Engine-Staging -> Main) (v777-15581079)

[CL 15628250 by max chen in Main branch]