Thin translucent support in the path tracer
Similar to rough glass support when using the translucent blend mode, the thin translucent lighting model blends in a thin (rough) glass model when Opacity is less than 1. When Ior is 1, plain transparency is used (again similarly to the solid glass case). Both cases account for the thin translucent color set by the shader to account for absorption through the slab.

This patch also refactors the GGX calls into a common place to minimize code duplication.

#rb Patrick.Kelly

#ushell-cherrypick of 15272638 by chris.kulla

#robomerge[STARSHIP] -Starship-RES

[CL 15660235 by chris kulla in Main branch]