Optimize the 'y' component evaluation of the Sobol' pattern
Since this dimension has a very simple Sobol matrix, we can generate it on the fly instead of loading it from the constant array.

Also add the option (enabled by default) to unroll the core Sobol generation loop. This was suggested by Brian Karis and seems to be faster for cases where execution if coherent. On the other hand, the path tracer seems faster overall with the "early" exit version (presumably because there is more divergence in that case).

#rb Brian.Karis

#ushell-cherrypick of 15651023 by chris.kulla

#robomerge[STARSHIP] -Starship-RES

[CL 15662620 by chris kulla in Main branch]