Fix culling of ISMs for ray tracing
Instanced static meshes were relying on clusters of nodes pre-computed in world-space for culling purposes. However, the data used to construct the clusters could become out of sync with the actual instance data. As a result, the culling could be highly inaccurate, depending on how the instance data was populated.

Temporarily remove the hierarchy, and replace it with a linear array of spheres in local space. Sort that data in the PerInstance structure, and update it on the render thread whenever instance data is updated.

Due to culling frequently getting disabled due to the prior implementation,  the linear array of bounds is generally faster than the previous hiearchy.

@ehart at Nvidia
#rb juan.canada

[CL 15686802 by Juan Canada in Main branch]