Normal offset contact shadows as well.
I previously killed the offset specifically for contact shadows to fix detail loss. At the time the amount of bias was larger than now and we still had the SMRT first sample leaking bug. Now that that is fixed and with the current 0.1 offset default which matches RT the difference of offset on and off on self shadow detail in Reverb is extremely small and not clearly better. It does though fix almost all of these triangle artifacts, at least on Echo.

[FYI] andrew.lauritzen

#ROBOMERGE-SOURCE: CL 15789996 in //UE5/Release-5.0-EarlyAccess/...
#ROBOMERGE-BOT: STARSHIP (Release-5.0-EarlyAccess -> Main) (v783-15756269)

[CL 15790848 by brian karis in ue5-main branch]