Add option for Coherent path tracing
When enabled, shuffle pixel coordinate and share random number streams to improve the coherence of execution. This can provide large speedups (2x or more) on some scenes, and even modest speedups on complex scenes (30% or so). However it comes at the price of somewhat distracting artifacts at low sample counts that may be objectionable.

Leaving this disabled by default for now to gather feedback.

#rb Juan.Canada

#ushell-cherrypick of 15825443 by chris.kulla

#robomerge[STARSHIP] -Starship-RES

[CL 15827145 by chris kulla in Main branch]