Optimized DrawStatsHUD by roughly 33% by reducing lookups in GetKerning.
GetKerning is currently the most expensive part of drawing/shaping a string for two reasons:
First, FSlateFontRenderer::GetKerning does a lookup to retrieve the FT_Face object, which requires computing the hash of the FFontData object. This is expensive because there's a filename string as part of the hash of FFontData, which is non-trivial to hash nor string-compare once we find the matching hash entry, and this is done for every character being drawn.
Second, once the FT_Face is retrieved, a monolithic map that is keyed on font face, kerning flags, size, scale, and glyph pair that contains the kerning values. This usually contains roughly a thousand entries, and adds up to a significant amount of time since the look-up is also done per character being drawn.

This change addresses the second issue by instead maintaining individual caches for each set of font parameters to reduce the number of elements that we have to search through for each character. Using this idea, we can sidestep the first problem by caching a pointer to the appropriate kerning cache that matches the parameters of the FCharacterListEntry, so that once the cache is warmed up we won't have to do FT_Face lookups by FFontData anymore. This results in a slight (50Kb) memory increase on editor builds as they are the only users of FCharacterList, but should reduce memory usage (-30KB) on shipping game builds as a result of the smaller kerning cache keys.

#rb [at]jamie.dale
[FYI] [at]andrew.ladenberger, [at]jordan.cristiano
#tests Made sure debug text still looks the same. Checked that glyph/character pairs that returned a kerning value prior to this change returned the same one after. Checked that the cache was indeed warmed up by seeing that we do not reach FSlateFontRenderer::GetKerningCache after "stat unit" has been rendering for a number of frames. Checked that the cache directory was full of entries after executing for a bit (actually revealed a bug due to bad MoveTemp).

#ROBOMERGE-SOURCE: CL 15826203 via CL 15827835 via CL 15828043 via CL 15828243
#ROBOMERGE-BOT: RELEASE (Release-Engine-Staging -> Main) (v785-15821902)
#ROBOMERGE[bot4]: -BOT4-Release-Engine-Staging

[CL 15828389 by evgenii babinets in Main branch]