Remove Message Dialog from Skeleton processing to fix a crash with destroying SAssetPicker while it's rendering thumbnails in it's Tick().
The repro was to simply drop DDC cache and open the asset picker filtered to all animation types using Shift+Alt+O. The creation of this dialog destroys transient UIs like SAssetPicker, which are destroyed  when their focus is lost. If these UIs have a thumbnail pool, as the SAssetPicker does, or any other owned object that subscribes to Tick(), then these could be inside Tick() when they are destroyed. This leads to general memory chaos policed by some ensures, and eventually an access violation as object lifetimes expire out of lockstep with each other.

The Dialog was created @2583044 which predates Load Errors Message logging, so workflow should be unhindered by removing this dialog.

#jira UE-110291
#rb francis.hurteau, matt.kuhlenschmidt

[CL 15828780 by geoff evans in Main branch]