Fix: the RHI validation layer assumed that UAV indices must be in the range 0-7.
This is true for D3D11, since SRVs and UAVs have different index spaces, but in Vulkan all the resources share the same space, so UAV indices can be larger than 7. The validation code overflowed the statically allocated UAV tracking buffer for shaders which happened to have UAVs in slots above 7. The solution is to dynamically size the tracking array, and reset it when a new shader or PSO is bound.

#rb none
#jira none
[FYI] Luke.Thatcher
#rnx

#ROBOMERGE-SOURCE: CL 15915732 in //UE5/Release-5.0-EarlyAccess/...
#ROBOMERGE-BOT: STARSHIP (Release-5.0-EarlyAccess -> Main) (v786-15839533)

[CL 15920148 by mihnea balta in ue5-main branch]