Revert computation of refracted Fresnel term for clear coat shading model of non-metallic surfaces.
While nested dielectrics have an Eta close to 1.0, and have a nearly unobstructed refracted Fresnel contribution, this is not true for a bottom interface with a separate normal, which can generate widly varying Fresnel values. This inconsistency amounts to a look regression for projects that have selected the "Clear Coat Enable Second Normal" project setting.

While reverting the behavior only for projects where this setting is enabled, carving out such a path for the bottom normal presents a modeling inconsistency (compared to original shading model without a varying bottom normal) in the case where the bottom normal and top normal are identical. Since we do not track IOR through interfaces, we cannot easily decouple modification of bottom-coat Fresnel due to Eta or normal variation. It is expected that this concept will be expressible in the future, but remains unsupported for now.

#rb juan.canada

#ushell-cherrypick of 15136790 by Patrick.Kelly

[CL 15942551 by Sebastien Hillaire in Main branch]