MoviePipeline: Refactored the way we detect ranges of time to render. Previously the range detection only worked for a sequence that was up to two levels deep (ie: either the top level sequence contained a camera cut track, or it contained a shot track which contained a camera cut track). Additionally the previous method did not take the accumulated sequence hierarchy into account which prevented shot tracks which have had their section's Time Scale property modified from producing the correct range to render.
In this change, the sequence evaluation tree is now walked to produce the ranges to render. This allows us to support arbitrarily deep sequence hierarchies as well as sequence hierarchies where some portions have been time scaled. As part of this change, shots are now expanded in isolation as part of the soloing process. This fixes issues where multiple shots would share some data (such as a sub-sequence) and when they all got expanded before rendering, the shared data would be over-expanded and thus result in incorrect renders.


#jira UE-103921, UE-105084, UE-107427, UE-107686
#rb Max.Chen

#ROBOMERGE-SOURCE: CL 15943508 via CL 15950656 via CL 15952047
#ROBOMERGE-BOT: RELEASE (Release-Engine-Staging -> Main) (v787-15839533)
#ROBOMERGE[bot4]: -BOT4-Release-Engine-Staging

[CL 15952051 by matt hoffman in Main branch]