UE-62702 - Clicking the button to remove a Data Table row while no row is selected causes the first row in the table to be selected
The various commands that depend on a valid selection (eg: Copy, Duplicate, etc) need to use a FCanExecuteAction delegate to ensure that a row is currently highlighted. For now, we can just use CanEditTable for that check.

The Jira only calls out the toolbar as having this issue, but we need to add FCanExecuteAction to the commands as well. This ensures that these commands are disabled if they're ever added to a menu in the future.

#jira UE-62702
#rb michael.noland

[CL 15989996 by dave jones2 in ue5-main branch]