UE-112805 - Crash with BP action menu
The original issue can't be reproduced, but it seems like the user is deleting a variable and then possibly right-clicking its old location before the UI had a chance to refresh. This would cause GetProperty() to return null, and cause a crash. Other actions in this function seem to already guard against this scenario, so we can just mimic that here.

Additionally, added a couple of extra checks for other actions where the pointer isn't checked before dereferencing.

#jira UE-112805
#review @michael.noland

[CL 15990736 by dave jones2 in ue5-main branch]