Add light culling for path tracer
Split the light list info finite (point, spot, rect) and infinite (skydome, directional) sets. Build a 2d grid of which finite lights overlap particular regions. The grid is aligned to the bounding box of all lights and collapsed along the shortext axis so that memory usage scales quadratically with resolution instead of cubically. The light list per cell is currently pre-allocated up to a maximum count per cell (overflow is simply cut and lights are dropped as they were prior to this change). In most scenes I have tried so far this maximum is never reached with the current defaults of 256^2 resolution and 128 max lights per cluster.

#rb Patrick.Kelly
#preflight 60785d289e2d7a00016e3032
#jira none

#ushell-cherrypick of 16022731 by chris.kulla

#robomerge[STARSHIP] -Starship-RES

[CL 16037096 by chris kulla in 4.27 branch]