Sequencer: Handle base post-process setting values when converting CameraAnims to TemplateSequences.
CameraAnims have the concept of a "base post-process setting" which contains a bunch of potentially non-default values
setup on the camera, which the animation relies on. These are now taken into account, and converted into Sequencer tracks
with only a default value on them. Previously, they were ignored, so the resulting TemplateSequence was incomplete.

#rb max.chen


#ROBOMERGE-SOURCE: CL 16042024 via CL 16042027
#ROBOMERGE-BOT: (v789-15992632)
#ROBOMERGE[bot4]: -BOT4-Release-Engine-Staging

[CL 16042031 by ludovic chabant in 4.27 branch]