USDImporter: fix for only one texture being imported from a USDz package
When importing from a USDz package, the resolved asset paths to textures
identify textures inside the package rather than distinct files on disk. Attempting
to hash files at these paths results in an empty hash since there is no such file at
that path. This causes the first texture imported to be cached with an empty
hash, and then every subsequent texture is similarly computed to have an empty
hash, yielding the first texture again when looking up into the cache.

Instead, we compute the hashes for textures inside USDz packages using the raw
texture data inside the package rather than trying to hash the contents of a file on
disk.

#jira UE-111979
#rb jeanmichel.dignard, daniel.coelho, jason.walter

[CL 16080981 by matt johnson in 4.27 branch]