[TraceLog] Remove callstack worker thread
Removing callstack tracing worker thread. Profiling shows that the performance impact is negligable. Additionally it solves some deadlock issues and makes the code simpler. In order to only trace unique callstack we track emitted ones in a TSherwoodSet. The insertion performance for this collection is very efficient (see https://probablydance.com/2017/02/26/i-wrote-the-fastest-hashtable/) which means offloading it to a worker thread was not giving the expected gains.

#rb martin.ridgers
#rnx

[CL 16089889 by Johan Berg in ue5-main branch]