Path Tracer light sampling functions only need 2 random numbers
Tweak the API so the light sampling functions only get 2 random numbers. This ensures we are consistent in which dimensions we use across all lights and make it easier to play with which dimensions are used for what. I played a bit with swizzling around the light selection vs light sampling dimensions but couldn't get consistent improvements, so I'm leaving it alone for now.

#rb Patrick.Kelly
#jira none

#ushell-cherrypick of 16079614 by chris.kulla


#ROBOMERGE-SOURCE: CL 16163463 in //UE4/Release-4.27/...
#ROBOMERGE-BOT: RELEASE (Release-4.27 -> Main) (v789-15992632)
#ROBOMERGE[bot4]: -Starship-RES

[CL 16163468 by chris kulla in Main branch]