UE-107119 - Crash in FKismetFunctionDragDropAction::DroppedOnPanel
Changed the GetHoveredGraph check to a soft-fail.

WM_MOUSEMOVE messages can arrive with the latest cursor coordinates, which can be in the "My Blueprint" pane. When this happen, the action's graph is set null. However, we can also end up processing WM_LBUTTONUP in the same tick, with cursor coordinates that are still in the graph pane. When this happens, we'll call DroppedOnPin, but the graph pointer will be null.

Null-checking should be safe here since we'll essentially ignore the action, which is then disposed of.

#jira UE-107119
#rb ben.zeigler

[CL 16186253 by dave jones2 in ue5-main branch]