Fix: the RHI validation layer assumed that UAV indices must be in the range 0-7.
This is true for D3D11, since SRVs and UAVs have different index spaces, but in Vulkan all the resources share the same space, so UAV indices can be larger than 7. The validation code overflowed the statically allocated UAV tracking buffer for shaders which happened to have UAVs in slots above 7. The solution is to dynamically size the tracking array, and reset it when a new shader or PSO is bound.

#rb Luke.Thatcher
#jira none
#rnx

#p4v-cherrypick 15920148

#ROBOMERGE-SOURCE: CL 16218895 in //UE4/Release-4.27/...
#ROBOMERGE-BOT: RELEASE (Release-4.27 -> Main) (v796-16191650)

[CL 16218907 by mihnea balta in Main branch]