DDC: Added UE::DerivedData::IBuildFunction and related types
IBuildFunction and the associated build context types define the interface for new derived data build functions. The inputs to a build function are referenced by FBuildDefinition. This only includes the functionality required to implement and register a build function, and does not include the functionality required to execute a build starting from a build definition.

#rb Zousar.Shaker
#rnx

#ROBOMERGE-SOURCE: CL 16278733 in //UE5/Main/...
#ROBOMERGE-BOT: STARSHIP (Main -> Release-Engine-Test) (v799-16237190)

[CL 16279132 by devin doucette in ue5-release-engine-test branch]