Integrate CL 16308862 from Dev-EngineMerge: PerfReportTool 4.47 - ExtraLinks summary Change 16308862 by ben.woodhouse[at]ROBOMERGE_FORTNITE_Dev_EngineMerge on 2021/05/12 22:00:30
PerfReportTool 4.47 - ExtraLinks summary
Adds a new ExtraLinks summary type, which reads a <csvfilename>.links file. The file contains lines with additional links in the format:
LongName,ShortName,URL

These links are added to a new Links section of the report. They are also added to summary table row, with the LongName as the key.

The section can be added to all reports by adding with -extraLinksSummary, or add the summary type "ExtraLinks" to the Report type XML.


#ROBOMERGE-SOURCE: CL 16308651 via CL 16309102
#ROBOMERGE-BOT: (v804-16311228)
#ROBOMERGE[bot4]: ue5-main

[CL 16315492 by ben woodhouse in ue5-release-engine-staging branch]